Move the tests behind the integration feature flag
